qemu/kvm: dmidecode: look for 'KVM'
We can check the product name for 'KVM' to determine if we're
accelerated. This only works for arm/aarch64 targets boot with
AAVMF and on RHEL x86 targets (which already have KVM in the cpuid
leaf), but it doesn't hurt to try on all.

(cherry picked from commit 6fe2ebaec32d8f554749807bc21983506647a9d5)