tests: Fix tests when run on AWS.
When running on AWS two of the tests failed.  This was because the
tests did _not_ include a working dmidecode binary, hence the platform
dmidecode binary was being run instead, and that was detecting as AWS
owing to the previous commit.  Fix this by supplying a working
dmidecode binary for those tests.